+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.23.0
+
+* Added `@pattern` function for pixel patterns drawing.
+  ```css
+  background: @pattern(
+    grid: 72;
+    match((int(x+y)%2) == 0) {
+      fill: #000;
+    }
+  )
+  ```
+<br /> <br />
+
+
+
 ## 0.22.0
 
 * Now use WebGL2 and GLSL3.0 for shader programs.